Major refactor in polyline tool to use a base class for other operators. Polyline Wall and Measure tool now share most of the logic by inheriting from a base class PolylineOperator. This should facilitate the implementation of other tools that use polyline.
