+++ src/bonsai/bonsai/bim/module/model/wall.py
@@ -38,6 +38,7 @@ from math import pi, sin, cos, degrees
 from mathutils import Vector, Matrix
 from bonsai.bim.module.model.opening import FilledOpeningGenerator
 from bonsai.bim.module.model.decorator import PolylineDecorator
+from bonsai.bim.module.model.polyline import PolylineOperator
 from typing import Optional
 from lark import Lark, Transformer
 
@@ -280,7 +281,7 @@ def recalculate_dumb_wall_origin(wall, new_origin=None):
         child.matrix_parent_inverse = wall.matrix_world.inverted()
 
 
-class DrawPolylineWall(bpy.types.Operator):
+class DrawPolylineWall(bpy.types.Operator, PolylineOperator):
     bl_idname = "bim.draw_polyline_wall"
     bl_label = "Draw Polyline Wall"
     bl_options = {"REGISTER", "UNDO"}
@@ -290,68 +291,7 @@ class DrawPolylineWall(bpy.types.Operator):
         return context.space_data.type == "VIEW_3D"
 
     def __init__(self):
-        self.mousemove_count = 0
-        self.action_count = 0
-        self.visible_objs = []
-        self.objs_2d_bbox = []
-        self.number_options = {
-            "0",
-            "1",
-            "2",
-            "3",
-            "4",
-            "5",
-            "6",
-            "7",
-            "8",
-            "9",
-            " ",
-            ".",
-            "+",
-            "-",
-            "*",
-            "/",
-            "'",
-            '"',
-            "=",
-        }
-        self.number_input = []
-        self.number_output = ""
-        self.number_is_negative = False
-        self.input_options = ["D", "A", "X", "Y"]
-        self.input_type = None
-        self.input_type = None
-        self.input_value_xy = [None, None]
-        self.input_ui = tool.Polyline.create_input_ui()
-        self.is_typing = False
-        self.snap_angle = None
-        self.snapping_points = []
-        self.instructions = """TAB: Cycle Input
-    M: Modify Snap Point
-    C: Close
-    Backspace: Remove
-    X Y: Axis
-    Shift: Lock axis
-"""
-        self.tool_state = tool.Polyline.create_tool_state()
-
-    def recalculate_inputs(self, context):
-        if self.number_input:
-            is_valid, self.number_output = tool.Polyline.validate_input(self.number_output, self.input_type)
-            self.input_ui.set_value(self.input_type, self.number_output)
-            if not is_valid:
-                self.report({"WARNING"}, "The number typed is not valid.")
-                return is_valid
-            else:
-                if self.input_type in {"X", "Y"}:
-                    tool.Polyline.calculate_distance_and_angle(context, self.input_ui, self.tool_state)
-                elif self.input_type in {"D", "A"}:
-                    tool.Polyline.calculate_x_y_and_z(context, self.input_ui, self.tool_state)
-                    tool.Polyline.calculate_distance_and_angle(context, self.input_ui, self.tool_state)
-                else:
-                    self.input_ui.set_value(self.input_type, self.number_output)
-            tool.Blender.update_viewport()
-            return is_valid
+        super().__init__()
 
     # TODO This is creating a hack in generate function from DumbWallGenerator
     # Come up with a better solution
@@ -380,131 +320,14 @@ class DrawPolylineWall(bpy.types.Operator):
                     DumbWallJoiner().join_V(wall1["obj"], wall2["obj"])
 
     def modal(self, context, event):
+        if event.type in {"MIDDLEMOUSE", "WHEELUPMOUSE", "WHEELDOWNMOUSE"}:
+            return {"PASS_THROUGH"}
 
-        if not self.tool_state.is_input_on:
-            if event.type == "MOUSEMOVE" or event.type == "INBETWEEN_MOUSEMOVE":
-                self.mousemove_count += 1
-                self.tool_state.mode = "Mouse"
-                self.tool_state.is_input_on = False
-                self.input_type = None
-                self.tool_state.input_type = None
-                PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-                tool.Snap.clear_snapping_ref()
-                tool.Blender.update_viewport()
-            else:
-                self.mousemove_count = 0
+        self.handle_mouse_move(context, event)
 
-            if self.mousemove_count == 2:
-                self.objs_2d_bbox = []
-                for obj in self.visible_objs:
-                    self.objs_2d_bbox.append(tool.Raycast.get_on_screen_2d_bounding_boxes(context, obj))
+        self.choose_axis(event)
 
-            if self.mousemove_count > 3:
-                detected_snaps = tool.Snap.detect_snapping_points(context, event, self.objs_2d_bbox, self.tool_state)
-                self.snapping_points = tool.Snap.select_snapping_points(context, event, self.tool_state, detected_snaps)
-                print(self.snapping_points)
-                tool.Polyline.calculate_distance_and_angle(context, self.input_ui, self.tool_state)
-                tool.Blender.update_viewport()
-                return {"RUNNING_MODAL"}
-
-            if event.value == "RELEASE" and event.type == "BACK_SPACE":
-                tool.Snap.remove_last_polyline_point()
-                tool.Blender.update_viewport()
-
-        if event.value == "RELEASE" and event.type == "LEFTMOUSE":
-            tool.Snap.insert_polyline_point(self.input_ui)
-            tool.Blender.update_viewport()
-
-        if event.value == "PRESS" and event.type == "X":
-            self.tool_state.axis_method = "X" if self.tool_state.axis_method != event.type else None
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
-
-        if event.value == "PRESS" and event.type == "Y":
-            self.tool_state.axis_method = "Y" if self.tool_state.axis_method != event.type else None
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
-
-        if event.value == "PRESS" and event.type == "C":
-            tool.Snap.close_polyline()
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
-
-        if self.tool_state.is_input_on and event.value == "PRESS" and event.type == "TAB":
-            self.recalculate_inputs(context)
-            index = self.input_options.index(self.input_type)
-            size = len(self.input_options)
-            self.input_type = self.input_options[((index + 1) % size)]
-            self.tool_state.input_type = self.input_options[((index + 1) % size)]
-            self.tool_state.mode = "Select"
-            self.is_typing = False
-            self.number_input = self.input_ui.get_formatted_value(self.input_type)
-            self.number_input = list(self.number_input)
-            self.number_output = "".join(self.number_input)
-            self.input_ui.set_value(self.input_type, self.number_output)
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
-
-        if not self.tool_state.is_input_on and event.value == "RELEASE" and event.type == "TAB":
-            self.recalculate_inputs(context)
-            self.tool_state.mode = "Select"
-            self.tool_state.is_input_on = True
-            self.input_type = "D"
-            self.tool_state.input_type = "D"
-            self.is_typing = False
-            self.number_input = self.input_ui.get_formatted_value(self.input_type)
-            self.number_input = list(self.number_input)
-            self.number_output = "".join(self.number_input)
-            self.input_ui.set_value(self.input_type, self.number_output)
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
-
-        if not self.tool_state.is_input_on and event.ascii in self.number_options:
-            self.recalculate_inputs(context)
-            self.tool_state.mode = "Edit"
-            self.tool_state.is_input_on = True
-            self.input_type = "D"
-            self.tool_state.input_type = "D"
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
-
-        if event.value == "RELEASE" and event.type in {"D", "A"}:
-            self.recalculate_inputs(context)
-            self.tool_state.mode = "Edit"
-            self.tool_state.is_input_on = True
-            self.input_type = event.type
-            self.tool_state.input_type = event.type
-            self.input_ui.set_value(self.input_type, "")
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
-
-        if self.input_type in self.input_options:
-            if (event.ascii in self.number_options) or (event.value == "RELEASE" and event.type == "BACK_SPACE"):
-                if not self.tool_state.mode == "Edit" and not (event.ascii == "=" or event.type == "BACK_SPACE"):
-                    self.number_input = []
-
-                if event.type == "BACK_SPACE":
-                    if len(self.number_input) <= 1:
-                        self.number_input = []
-                    else:
-                        self.number_input.pop(-1)
-                elif event.ascii == "=":
-                    if self.number_input[0] == "=":
-                        self.number_input.pop(0)
-                    else:
-                        self.number_input.insert(0, "=")
-                else:
-                    self.number_input.append(event.ascii)
-
-                if not self.number_input:
-                    self.number_output = "0"
-
-                self.tool_state.mode = "Edit"
-                self.is_typing = True
-                self.number_output = "".join(self.number_input)
-                self.input_ui.set_value(self.input_type, self.number_output)
-                PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-                tool.Blender.update_viewport()
+        self.handle_snap_selection(context, event)
 
         if not self.tool_state.is_input_on and event.value == "RELEASE" and event.type in {"RET", "NUMPAD_ENTER", "RIGHTMOUSE"}:
             self.create_walls_from_polyline(context)
@@ -513,72 +336,21 @@ class DrawPolylineWall(bpy.types.Operator):
             tool.Blender.update_viewport()
             return {"FINISHED"}
 
-        if self.tool_state.is_input_on and event.value == "RELEASE" and event.type in {"RET", "NUMPAD_ENTER", "RIGHTMOUSE"}:
-            is_valid = self.recalculate_inputs(context)
-            if is_valid:
-                tool.Snap.insert_polyline_point(self.input_ui)
-            self.tool_state.mode = "Mouse"
-            self.tool_state.is_input_on = False
-            self.input_type = None
-            self.tool_state.input_type = None
-            self.number_input = []
-            self.number_output = ""
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
+        self.handle_keyboard_input(context, event)
 
-        if event.value == "PRESS" and event.type == "M":
-            self.snapping_points = tool.Snap.modify_snapping_point_selection(self.snapping_points)
-            tool.Polyline.calculate_distance_and_angle(context, self.input_ui, self.tool_state)
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
+        self.handle_inserting_polyline(context, event)
 
-        if event.type in {"MIDDLEMOUSE", "WHEELUPMOUSE", "WHEELDOWNMOUSE"}:
-            return {"PASS_THROUGH"}
+        result = self.handle_cancelation(context, event)
+        if result is not None:
+            return result
 
-        if self.tool_state.is_input_on:
-            if event.value == "RELEASE" and event.type in {"ESC"}:
-                self.recalculate_inputs(context)
-                self.tool_state.mode = "Mouse"
-                self.tool_state.is_input_on = False
-                self.input_type = None
-                self.tool_state.input_type = None
-                PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-                tool.Blender.update_viewport()
-        else:
-            if event.value == "RELEASE" and event.type in {"ESC"}:
-                self.tool_state.axis_method = None
-                PolylineDecorator.uninstall()
-                tool.Snap.clear_polyline()
-                tool.Blender.update_viewport()
-                return {"CANCELLED"}
-
-        print(">>> ", self.tool_state.mode)
         return {"RUNNING_MODAL"}
 
     def invoke(self, context, event):
-        if context.space_data.type == "VIEW_3D":
-            PolylineDecorator.install(context)
-            tool.Snap.clear_snapping_point()
-
-            self.tool_state.use_default_container = True
-            self.tool_state.axis_method = None
-            self.tool_state.plane_method = "XY"
-            self.tool_state.mode = "Mouse"
-            tool.Snap.set_tool_state(self.tool_state)
-            self.visible_objs = tool.Raycast.get_visible_objects(context)
-            for obj in self.visible_objs:
-                self.objs_2d_bbox.append(tool.Raycast.get_on_screen_2d_bounding_boxes(context, obj))
-            detected_snaps = tool.Snap.detect_snapping_points(context, event, self.objs_2d_bbox, self.tool_state)
-            self.snapping_points = tool.Snap.select_snapping_points(context, event, self.tool_state, detected_snaps)
-            tool.Polyline.calculate_distance_and_angle(context, self.input_ui, self.tool_state)
-
-            PolylineDecorator.update(event, self.tool_state, self.input_ui, self.snapping_points[0])
-            tool.Blender.update_viewport()
-            context.window_manager.modal_handler_add(self)
-            return {"RUNNING_MODAL"}
-        else:
-            self.report({"WARNING"}, "Active space must be a View3d")
-            return {"CANCELLED"}
+        super().invoke(context, event)
+        self.tool_state.use_default_container = True
+        self.tool_state.plane_method = "XY"
+        return {"RUNNING_MODAL"}
 
 
 class DumbWallAligner:
